base windows with layout

+++ Assets/Scripts/Serialization/EditorJsonFileSaver.cs
@@ -1,5 +1,6 @@
 #if UNITY_EDITOR
 using System.IO;
+using UnityEngine;
 
 namespace JHI.Dict.Serialization
 {
@@ -10,10 +11,19 @@ namespace JHI.Dict.Serialization
 
         public static void SaveTestFile(string json)
         {
-            var path = Path.Combine(FILE_PATH, FILE_NAME);
+            if (string.IsNullOrEmpty(json))
+                return;
+            
+            var path = $"{Application.dataPath}{FILE_PATH}";
+            Debug.Log($"Save json to file {path}");
+            
             if (File.Exists(path))
                 File.Delete(path);
-            
+
+            if (!Directory.Exists(path))
+                Directory.CreateDirectory(path);
+
+            path += FILE_NAME;
             using (var sv = File.CreateText(path))
             {
                 sv.Write(json);
